tests structure refactor wip
