+++ lib/api/package.json
@@ -6,7 +6,7 @@
   "types": "types/index.d.ts",
   "scripts": {
     "build": "tslint --config ../../tslint.json --project tsconfig.json && ttsc",
-    "test": "for d in test/*/; do mocha ${d}/setup.ts ${d}/**/*.spec.ts -r ts-node/register -r tsconfig-paths/register || exit 1; done"
+    "test": "mocha test/setup.ts test/**/*.spec.ts -r ts-node/register -r tsconfig-paths/register || exit 1"
   },
   "keywords": [
     "extension",
@@ -50,6 +50,7 @@
     "mocha": "^5.1.1",
     "sinon": "^5.1.1",
     "sinon-chrome": "^2.3.2",
+    "ts-mockito": "^2.3.1",
     "ts-node": "^6.1.0",
     "tsconfig-paths": "^3.7.0",
     "tslint": "^5.11.0",
