+++ lib/api/src/browserAction/chrome/BrowserAction.ts
@@ -86,7 +86,7 @@ export class BrowserAction extends AbstractBrowserAction {
    */
   public registerEvents (dispatcher: Dispatcher) : void {
     chrome.browserAction.onClicked.addListener(({ id }) => {
-      dispatcher.fire(new BrowserActionClickedEvent({ id }))
+      dispatcher.fire(new BrowserActionClickedEvent(id))
     })
   }
 }
