--- lib/api/test/browserAction/spec/BrowserAction.spec.ts
@@ -1,9 +1,0 @@
-import { tests as ExtensionTests } from './ExtensionsBrowserAction.spec'
-import { tests as SafariTests } from './SafariBrowserAction.spec'
-import { tests as ChromeTests } from './ChromeBrowserAction.spec'
-
-describe('Browser Action', () => {
-  ChromeTests()
-  ExtensionTests()
-  SafariTests()
-})
