--- lib/api/test/messaging/safari.ts
@@ -1,13 +1,0 @@
-import * as sinon from 'sinon'
-
-export const safari = {
-  application: {
-    addEventListener: sinon.spy()
-  },
-
-  self: {
-    tab: {
-      dispatchMessage: sinon.spy()
-    }
-  }
-}
