--- lib/api/test/mocks/setup.ts
@@ -1,2 +1,0 @@
-/// <reference types="mocha" />
-/// <reference types="chai" />
