+++ lib/api/test/spec/browserAction/all.spec.ts
@@ -1,0 +1,9 @@
+import generate from '../../utils/generate'
+
+import chrome from './chrome'
+import extensions from './extensions'
+import safari from './safari'
+
+describe('BrowserAction', () => {
+  generate({ chrome, extensions, safari })
+})
